Move `CacheHooksService` logic to domain
