+++ src/domain/queues/queues-repository.interface.ts
@@ -1,18 +1,22 @@
 import { QueuesApiModule } from '@/datasources/queues/queues-api.module';
 import { QueuesRepository } from '@/domain/queues/queues-repository';
-import { Module, OnModuleInit } from '@nestjs/common';
-import { Event } from '@/routes/cache-hooks/entities/event.entity';
+import { Module } from '@nestjs/common';
+import { ConsumeMessage } from 'amqplib';
 
 export const IQueuesRepository = Symbol('IQueuesRepository');
 
-export interface IQueuesRepository extends OnModuleInit {
+export interface IQueuesRepository {
   /**
-   * Subscribes to {@link Event}s.
+   * Subscribes to messages from the specified queue.
    *
-   * @param onEvent - Callback to be executed when an {@link Event} is received.
-   * @returns {Promise<unknown>} A Promise that resolves when the subscription is successful.
+   * @param queueName - The name of the queue to subscribe to.
+   * @param fn - The callback function to be executed when a new message is received.
+   * @returns {Promise<void>} A Promise that resolves when the subscription is successful.
    */
-  onEvent(listener: (event: Event) => Promise<unknown>): void;
+  subscribe(
+    queueName: string,
+    fn: (msg: ConsumeMessage) => Promise<void>,
+  ): Promise<void>;
 }
 
 @Module({
